[selection] move sort-selected to experimental
